test: add comments and rename test for timer robustness
The name of the test did not make it clear what it was about. (It also
used "timer" in the name instead of "timers" like all the other tests.)
I also added a comment to be extra clear about the test purpose and a
link to the issue that was originally filed about it.

PR-URL: https://github.com/nodejs/node/pull/54008
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Chengzhong Wu <legendecas@gmail.com>
Reviewed-By: Trivikram Kamat <trivikr.dev@gmail.com>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Jake Yuesong Li <jake.yuesong@gmail.com>